Version 1.2: improved performance and optional GC
Version 1.2 features:
- Added a --GC (-gc) flag to make GC calculation optional.
- Added functionality for the --output (-o) flag, which is now required. It specifies the directory where your report will appear.
- Improved N50 calculation, which is now MUCH faster (details below).

In the first two versions, the script would make a list of all sequences, sort them in order and then make a very long string to determine the length, and use a similar method to calculate the N50. However, storing such a long string is not very nice for the computer, so now it simply stores a number corresponding to the length and the index.